cli/requirements.txt
 openai
 click
 sqlalchemy
 python-dotenv
 gitpython